updated next & react to latest versions, changed revalidation logic for data and isr caching
