+++ package.json
@@ -18,9 +18,9 @@
     "@vercel/speed-insights": "^1.0.14",
     "clsx": "^2.1.1",
     "motion": "^12.0.3",
-    "next": "^14.2.16",
-    "react": "^18",
-    "react-dom": "^18",
+    "next": "^15.1.6",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-infinite-scroller": "^1.2.6"
   },
   "devDependencies": {
@@ -28,7 +28,7 @@
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "eslint": "^8",
-    "eslint-config-next": "14.2.15",
+    "eslint-config-next": "^15.1.6",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
