+++ src/components/SeriesItem.tsx
@@ -6,9 +6,8 @@ import { useState } from "react";
 import { ChevronDownIcon } from "@heroicons/react/24/solid";
 import clsx from "clsx";
 import { motion } from "motion/react";
-import { getSlugFromSetId } from "@/utilities/slugs";
+import { getSlugFromSetId, seriesNameToSlug } from "@/utilities/slugs";
 import { ArrowRightIcon } from "@heroicons/react/24/outline";
-import { seriesNameToSlug } from "@/utilities/data";
 
 export default function SeriesItem({
   name,
