+++ src/components/SetCards.tsx
@@ -22,10 +22,10 @@ import SortSelect from "./SortSelect";
 import FlyoutMenu from "./FlyoutMenu";
 import { BinderView } from "./BinderView";
 import { motion } from "motion/react";
-import { combineRarities, formatPrice, getPrice } from "@/utilities/data";
 import Image from "next/image";
 import { Card, Rarities, Set, Subset } from "@/utilities/types";
 import InfiniteScroll from "react-infinite-scroller";
+import { combineRarities, formatPrice, getPrice } from "@/utilities/formatting";
 
 const cardReducer = (state: any, action: any) => {
   // console.log({ state, action });
